tests/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-dotenv>=1.0.0
supabase>=2.0.0
faker>=19.0.0
pytest-html>=4.0.0
pytest-cov>=4.1.0
